+++ README.md
@@ -34,8 +34,8 @@
   - [x] 회원 가입시 받는 정보는 `email`, `name`, `password`이다.
 - [x] 유저는 로그인 할 수 있다.
   - [x] 로그인하고 나면 로그인 버튼은 로그아웃 버튼으로 변경되어야 한다.
-- [ ] 유저는 로그아웃할 수 있다.
-  - [ ] 로그아웃하면 `/` 루트 페이지로 돌아온다.
+- [x] 유저는 로그아웃할 수 있다.
+  - [x] 로그아웃하면 `/` 루트 페이지로 돌아온다.
 
 ## 🎯🎯 step2
