+++ docs/README.md
@@ -15,8 +15,8 @@
     - [x] input 하단의 에러 메시지 렌더링
 - [x] 로그인하고 나면 로그인 버튼은 로그아웃 버튼으로 변경 된다.
 - [x] 로그인한 유저만 모든 기능을 조회할 수 있다.
-- [ ] 유저는 로그아웃 할 수 있다.
-  - [ ] 로그아웃시 `/` 루트 페이지로 돌아 온다.
+- [x] 유저는 로그아웃 할 수 있다.
+  - [x] 로그아웃시 `/` 루트 페이지로 돌아 온다.
 
 ## 테스트 케이스
 
@@ -28,4 +28,4 @@
   - [x] password는 최소 6자이상, 영숫자 및 특수문자의 조합으로 이루어져야 한다. (TDD)
 - [x] 로그인 실패 시, 에러 메시지를 렌더링한다.
 - [x] 로그인 성공 시, 메뉴 버튼들과 로그아웃 버튼이 화면에 노출되고 메인 페이지로 이동한다.
-- [ ] 로그아웃 성공 시, 로그인 버튼이 화면에 노출되고 메인 페이지로 이동한다.
+- [x] 로그아웃 성공 시, 로그인 버튼이 화면에 노출되고 메인 페이지로 이동한다.
