+++ docs/index.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="kr">
+  <head>
+    <meta charset="UTF-8" />
+    <title>🚇 지하철 노선도</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <script defer src="bundle.js"></script></head>
+  <body>
+    <div id="app">
+      <div class="d-flex justify-center mt-5 w-100">
+        <div class="w-100">
+          <header class="my-4">
+              <h1 data-link="/" class="js-link text-center font-bold cursor-pointer">🚇 지하철 노선도</h1>
+          </header>
+          <nav class="d-flex justify-center">
+            <div id="menu-buttons-container" class="mx-1"></div>
+            <div id="sign-button-container"></div>
+          </nav>
+          <main id="main-container" class="mt-10 d-flex justify-center"></main>
+        </div>
+      </div>
+    </div>
+  </body>
+</html>
