+++ src/js/subway/utils/index.js
@@ -1,2 +1,3 @@
 export { isValidEmail, isValidName, isValidPassword, findInValidInput } from './validate';
 export { getUserName } from './getUserName';
+export { routeTo, getRedirectedPath } from './route';
